fixup! security enhancements and move to shared helper
